apps/frontend/requirements.txt
# Core Streamlit
streamlit==1.28.1
streamlit-option-menu==0.3.6
streamlit-chat==0.1.1
streamlit-authenticator==0.2.3

# Data Visualization
plotly==5.17.0
pandas==2.1.3
numpy==1.26.2

# HTTP Client
httpx==0.25.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2


